fixed short handling

+++ vertx-pojo-mapper-json/src/main/java/de/braintags/io/vertx/pojomapper/json/typehandler/handler/ShortTypeHandler.java
@@ -58,7 +58,7 @@ public class ShortTypeHandler extends AbstractTypeHandler {
    */
   @Override
   public void intoStore(Object source, IField field, Handler<AsyncResult<ITypeHandlerResult>> resultHandler) {
-    success(source == null ? source : ((Short) source).shortValue(), resultHandler);
+    success(source == null ? source : ((Short) source).intValue(), resultHandler);
   }
 
 }
